chore(logs): logs redux-toolkit refactor
