+++ packages/blockchain-wallet-v4-frontend/src/data/actions.ts
@@ -10,7 +10,7 @@ import * as components from './components/actions'
 import * as custodial from './custodial/actions'
 import * as form from './form/actions'
 import * as goals from './goals/actions'
-import * as logs from './logs/actions'
+import { actions as logs } from './logs/slice'
 import * as middleware from './middleware/actions'
 import * as ws from './middleware/webSocket/coins/actions'
 import * as modals from './modals/actions'
