--- packages/blockchain-wallet-v4-frontend/src/data/logs/actionTypes.js
@@ -1,3 +1,0 @@
-export const LOG_INFO_MSG = 'LOG_INFO_MSG'
-export const LOG_ERROR_MSG = 'LOG_ERROR_MSG'
-export const SET_LOG_LEVEL = 'SET_LOG_LEVEL'
