+++ packages/blockchain-wallet-v4-frontend/src/data/rootReducer.ts
@@ -10,7 +10,8 @@ import components from './components/reducers'
 import { custodialReducer as custodial } from './custodial/reducers'
 import form from './form/reducers'
 import goals from './goals/reducers'
-import logs from './logs/reducers'
+// import logs from './logs/reducers'
+import { logsReducer as logs } from './logs/slice.js'
 import { modalsReducer } from './modals/reducers'
 import profile from './modules/profile/reducers'
 import rates from './modules/rates/reducers'
