+++ packages/blockchain-wallet-v4-frontend/src/middleware/autoDisconnection.js
@@ -59,7 +59,7 @@ const blackListedActivityTypes = [
   actionTypes.modules.profile.FETCH_USER_DATA_SUCCESS,
   actionTypes.modules.profile.FETCH_TIERS_SUCCESS,
   // LOGS
-  actionTypes.logs.LOG_ERROR_MSG,
+  // actionTypes.logs.LOG_ERROR_MSG,
   // FORMS
   '@@redux-form/CLEAR_SUBMIT_ERRORS',
   '@@redux-form/STOP_ASYNC_VALIDATION',
